precipitation_forecasting/requirements.txt
# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep Learning
torch>=1.12.0
tensorflow>=2.9.0
keras>=2.9.0

# Time Series
statsmodels>=0.13.0
prophet>=1.1.0
sktime>=0.13.0

# Meteorological Data
cdsapi>=0.5.1
xarray>=0.19.0
netcdf4>=1.5.8
cfgrib>=0.9.0
eccodes>=1.4.0

# Geospatial
cartopy>=0.20.0
rasterio>=1.2.0
geopandas>=0.10.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0
joblib>=1.1.0

# Development
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
jupyter>=1.0.0